+++ lib/src/events/restrr_event.dart
@@ -1,6 +1,6 @@
 import '../../restrr.dart';
 
-class RestrrEvent {
+abstract class RestrrEvent {
   final Restrr api;
 
   const RestrrEvent({required this.api});