+++ lib/src/events/self_user_refresh_event.dart
@@ -1,8 +1,0 @@
-
-import '../../restrr.dart';
-
-class SelfUserRefreshEvent extends RestrrEvent {
-  final User selfUser;
-
-  const SelfUserRefreshEvent({required super.api, required this.selfUser});
-}